refactor: UserInfo -> member 패키지로 이동
