+++ src/main/java/roomescape/auth/member/UserInfo.java
@@ -1,4 +1,4 @@
-package roomescape.auth;
+package roomescape.auth.member;
 
 import roomescape.domain.member.Role;
